MS1 Master

+++ app/imports/api/departments/methods.js
@@ -4,7 +4,8 @@ import { SimpleSchema } from 'meteor/aldeed:simple-schema';
 import { DepartmentSchema } from './department-schema';
 import DepartmentService from './department-service.js';
 import { IdSchema } from '../schemas.js';
-import { checkUserId } from '../checkers.js';
+import { UserRoles } from '../constants';
+
 
 export const insert = new ValidatedMethod({
   name: 'Departments.insert',
@@ -12,9 +13,21 @@ export const insert = new ValidatedMethod({
   validate: DepartmentSchema.validator(),
 
   run(doc) {
-    checkUserId(
-      this.userId, 'Unauthorized user cannot create a department'
-    );
+    if (!this.userId) {
+      throw new Meteor.Error(
+        403, 'Unauthorized user cannot create a department'
+      );
+    }
+    
+    const { organizationId } = doc;
+    const canEditOrgSettings = Roles.userIsInRole(this.userId, UserRoles.CHANGE_ORG_SETTINGS, organizationId);
+
+    if (!canEditOrgSettings) {
+      throw new Meteor.Error(
+        403,
+        'User is not authorized for editing organization settings'
+      );
+    }
 
     return DepartmentService.insert(doc);
   }
@@ -28,9 +41,21 @@ export const update = new ValidatedMethod({
   }]).validator(),
 
   run(doc) {
-    checkUserId(
-      this.userId, 'Unauthorized user cannot update a department'
-    );
+    if (!this.userId) {
+      throw new Meteor.Error(
+        403, 'Unauthorized user cannot update a department'
+      );
+    }
+
+    const { organizationId } = doc;
+    const canEditOrgSettings = Roles.userIsInRole(this.userId, UserRoles.CHANGE_ORG_SETTINGS, organizationId);
+
+    if (!canEditOrgSettings) {
+      throw new Meteor.Error(
+        403,
+        'User is not authorized for editing organization settings'
+      );
+    }
 
     return DepartmentService.update(doc);
   }
@@ -42,9 +67,21 @@ export const remove = new ValidatedMethod({
   validate: IdSchema.validator(),
 
   run(doc) {
-    checkUserId(
-      this.userId, 'Unauthorized user cannot remove a department'
-    );
+    if (!this.userId) {
+      throw new Meteor.Error(
+        403, 'Unauthorized user cannot remove a department'
+      );
+    }
+
+    const { organizationId } = doc;
+    const canEditOrgSettings = Roles.userIsInRole(this.userId, UserRoles.CHANGE_ORG_SETTINGS, organizationId);
+
+    if (!canEditOrgSettings) {
+      throw new Meteor.Error(
+        403,
+        'User is not authorized for editing organization settings'
+      );
+    }
 
     return DepartmentService.remove(doc);
   }
