+++ app/imports/api/standards-book-sections/methods.js
@@ -7,7 +7,8 @@ import {
   StandardsBookSectionSchema
 } from './standards-book-section-schema.js';
 import { IdSchema } from '../schemas.js';
-import { checkUserId } from '../checkers.js';
+import { UserRoles } from '../constants';
+
 
 export const insert = new ValidatedMethod({
   name: 'StandardsBookSections.insert',
@@ -15,10 +16,21 @@ export const insert = new ValidatedMethod({
   validate: StandardsBookSectionSchema.validator(),
 
   run(doc) {
-    checkUserId(
-      this.userId,
-      'Unauthorized user cannot create a standards book section'
-    );
+    if (!this.userId) {
+      throw new Meteor.Error(
+        403, 'Unauthorized user cannot create a standards book section'
+      );
+    }
+
+    const { organizationId } = doc;
+    const canEditOrgSettings = Roles.userIsInRole(this.userId, UserRoles.CHANGE_ORG_SETTINGS, organizationId);
+
+    if (!canEditOrgSettings) {
+      throw new Meteor.Error(
+        403,
+        'User is not authorized for editing organization settings'
+      );
+    }
 
     return StandardsBookSectionService.insert(doc);
   }
@@ -33,10 +45,21 @@ export const update = new ValidatedMethod({
   ]).validator(),
 
   run(doc) {
-    checkUserId(
-      this.userId,
-      'Unauthorized user cannot update a standards book section'
-    );
+    if (!this.userId) {
+      throw new Meteor.Error(
+        403, 'Unauthorized user cannot update a standards book section'
+      );
+    }
+
+    const { organizationId } = doc;
+    const canEditOrgSettings = Roles.userIsInRole(this.userId, UserRoles.CHANGE_ORG_SETTINGS, organizationId);
+
+    if (!canEditOrgSettings) {
+      throw new Meteor.Error(
+        403,
+        'User is not authorized for editing organization settings'
+      );
+    }
 
     return StandardsBookSectionService.update(doc);
   }
@@ -48,10 +71,21 @@ export const remove = new ValidatedMethod({
   validate: IdSchema.validator(),
 
   run(doc) {
-    checkUserId(
-      this.userId,
-      'Unauthorized user cannot remove a standards book section'
-    );
+    if (!this.userId) {
+      throw new Meteor.Error(
+        403, 'Unauthorized user cannot remove a standards book section'
+      );
+    }
+
+    const { organizationId } = doc;
+    const canEditOrgSettings = Roles.userIsInRole(this.userId, UserRoles.CHANGE_ORG_SETTINGS, organizationId);
+
+    if (!canEditOrgSettings) {
+      throw new Meteor.Error(
+        403,
+        'User is not authorized for editing organization settings'
+      );
+    }
 
     return StandardsBookSectionService.remove(doc);
   }
