+++ app/imports/api/standards-types/methods.js
@@ -4,24 +4,37 @@ import { SimpleSchema } from 'meteor/aldeed:simple-schema';
 import StandardsTypeService from './standards-type-service.js';
 import { StandardsTypeSchema } from './standards-type-schema.js';
 import { IdSchema } from '../schemas.js';
-import { checkUserId } from '../checkers.js';
+import { UserRoles } from '../constants';
+
 
 export const insert = new ValidatedMethod({
-  name: 'StandardsTypes.insert',
+  name: 'StandardTypes.insert',
 
   validate: StandardsTypeSchema.validator(),
 
   run(doc) {
-    checkUserId(
-      this.userId, 'Unauthorized user cannot create a standard type'
-    );
+    if (!this.userId) {
+      throw new Meteor.Error(
+        403, 'Unauthorized user cannot create a standard type'
+      );
+    }
+
+    const { organizationId } = doc;
+    const canEditOrgSettings = Roles.userIsInRole(this.userId, UserRoles.CHANGE_ORG_SETTINGS, organizationId);
+
+    if (!canEditOrgSettings) {
+      throw new Meteor.Error(
+        403,
+        'User is not authorized for editing organization settings'
+      );
+    }
 
     return StandardsTypeService.insert(doc);
   }
 });
 
 export const update = new ValidatedMethod({
-  name: 'StandardsTypes.update',
+  name: 'StandardTypes.update',
 
   validate: new SimpleSchema([IdSchema, {
     name: { type: String },
@@ -29,23 +42,47 @@ export const update = new ValidatedMethod({
   }]).validator(),
 
   run(doc) {
-    checkUserId(
-      this.userId, 'Unauthorized user cannot update a standard type'
-    );
+    if (!this.userId) {
+      throw new Meteor.Error(
+        403, 'Unauthorized user cannot update a standard type'
+      );
+    }
+
+    const { organizationId } = doc;
+    const canEditOrgSettings = Roles.userIsInRole(this.userId, UserRoles.CHANGE_ORG_SETTINGS, organizationId);
+
+    if (!canEditOrgSettings) {
+      throw new Meteor.Error(
+        403,
+        'User is not authorized for editing organization settings'
+      );
+    }
 
     return StandardsTypeService.update(doc);
   }
 });
 
 export const remove = new ValidatedMethod({
-  name: 'StandardsTypes.remove',
+  name: 'StandardTypes.remove',
 
   validate: IdSchema.validator(),
 
   run(doc) {
-    checkUserId(
-      this.userId, 'Unauthorized user cannot remove a standard type'
-    );
+    if (!this.userId) {
+      throw new Meteor.Error(
+        403, 'Unauthorized user cannot remove a standard type'
+      );
+    }
+
+    const { organizationId } = doc;
+    const canEditOrgSettings = Roles.userIsInRole(this.userId, UserRoles.CHANGE_ORG_SETTINGS, organizationId);
+
+    if (!canEditOrgSettings) {
+      throw new Meteor.Error(
+        403,
+        'User is not authorized for editing organization settings'
+      );
+    }
 
     return StandardsTypeService.remove(doc);
   }
